Converts connector list to patternfly table
