+++ ui/packages/ui/package.json
@@ -21,13 +21,14 @@
     "@types/enzyme": "^3.9.0",
     "@types/enzyme-adapter-react-16": "^1.0.5",
     "@types/jest": "^24.0.9",
+    "@types/lodash": "^4.14.161",
     "@types/node": "^11.9.6",
     "@types/react": "^16.8.6",
     "@types/react-dom": "^16.8.2",
-    "@types/lodash": "^4.14.161",
     "@types/react-router": "^5.1.8",
     "@types/react-router-dom": "^5.1.5",
     "@types/yup": "^0.29.6",
+    "autoprefixer": "^9.8.6",
     "copy-webpack-plugin": "^5.0.4",
     "css-loader": "^4.3.0",
     "enzyme": "^3.7.0",
@@ -40,11 +41,15 @@
     "keycloak-js": "9.0.0",
     "mini-css-extract-plugin": "^1.0.0",
     "optimize-css-assets-webpack-plugin": "^5.0.4",
+    "postcss-css-variables": "^0.17.0",
+    "postcss-loader": "^3.0.0",
+    "postcss-nested": "^4.2.3",
     "prettier": "^1.15.2",
     "prop-types": "^15.6.1",
     "raw-loader": "^4.0.2",
     "regenerator-runtime": "^0.13.1",
     "rimraf": "^2.6.2",
+    "serve": "11.3.2",
     "style-loader": "^2.0.0",
     "svg-url-loader": "^6.0.0",
     "ts-jest": "^24.0.0",
@@ -60,17 +65,14 @@
     "webpack-bundle-analyzer": "^3.9.0",
     "webpack-cli": "^4.0.0",
     "webpack-dev-server": "^3.11.0",
-    "webpack-merge": "^5.2.0",
-    "serve": "11.3.2",
-    "autoprefixer": "^9.8.6",
-    "postcss-loader": "^3.0.0",
-    "postcss-nested": "^4.2.3"
+    "webpack-merge": "^5.2.0"
   },
   "dependencies": {
     "@debezium/ui-models": "^1.0.0",
     "@debezium/ui-services": "^1.0.0",
     "@patternfly/react-core": "^4.75.2",
     "@patternfly/react-icons": "^4.7.16",
+    "@patternfly/react-table": "^4.19.24",
     "formik": "^2.1.5",
     "lodash": "^4.17.20",
     "react": "16.13.1",
