+++ ui/packages/ui/src/app/pages/connectors/ConnectorsTableComponent.tsx
@@ -2,7 +2,6 @@ import { Connector } from "@debezium/ui-models";
 import { Services } from "@debezium/ui-services";
 import {
   Button,
-  DataList,
   EmptyState,
   EmptyStateBody,
   EmptyStateIcon,
@@ -12,24 +11,16 @@ import {
   Title
 } from "@patternfly/react-core";
 import { CubesIcon } from "@patternfly/react-icons";
+import { cellWidth, Table, TableBody, TableHeader } from "@patternfly/react-table";
 import React from "react";
-import { PageLoader } from "src/app/components";
-import { ApiError, fetch_retry } from "src/app/shared";
+import { PageLoader, ToastAlertComponent } from "src/app/components";
+import { AppLayoutContext } from 'src/app/Layout/AppLayoutContext';
+import { ApiError, ConfirmationButtonStyle, ConfirmationDialog, ConfirmationType, ConnectorTypeId, fetch_retry } from "src/app/shared";
 import { WithLoader } from "src/app/shared/WithLoader";
-import { ConnectorListItem } from "./ConnectorListItem";
+import { ConnectorIcon } from './ConnectorIcon';
 import "./ConnectorsTableComponent.css";
-
-/**
- * Sorts the connectors by name.
- * @param connectors
- */
-function getSortedConnectors(connectors: Connector[]) {
-  const sortedConns: Connector[] = connectors.sort((thisConn, thatConn) => {
-    return thisConn.name.localeCompare(thatConn.name);
-  });
-
-  return sortedConns;
-}
+import { ConnectorStatus } from './ConnectorStatus';
+import { ConnectorTask } from './ConnectorTask';
 
 type ICreateConnectorCallbackFn = (connectorNames: string[], clusterId: number) => void
 
@@ -41,10 +32,54 @@ interface IConnectorsTableComponentProps {
 
 export const ConnectorsTableComponent: React.FunctionComponent<IConnectorsTableComponentProps> = (props: IConnectorsTableComponentProps) => {
   const [connectors, setConnectors] = React.useState<Connector[]>([] as Connector[]);
+  const [tableRows, setTableRows] = React.useState<any[]>([]);
 
   const [loading, setLoading] = React.useState(true);
   const [apiError, setApiError] = React.useState<boolean>(false);
   const [errorMsg, setErrorMsg] = React.useState<Error>(new Error());
+
+  const appLayoutContext = React.useContext(AppLayoutContext);
+  const [showDeleteDialog, setShowDeleteDialog] = React.useState(false);
+  const [alerts, setAlerts] = React.useState<any[]>([]);
+  const [connectorToDelete, setConnectorToDelete] = React.useState('');
+
+  const addAlert = (type: string, heading: string, msg?: string) => {
+    const alertsCopy = [...alerts];
+    const uId = new Date().getTime();
+    const newAlert = {
+      title: heading,
+      variant: type,
+      key: uId,
+      message: msg ? msg : "",
+    };
+    alertsCopy.push(newAlert);
+    setAlerts(alertsCopy);
+  };
+
+  const removeAlert = (key: string) => {
+    setAlerts([...alerts.filter((el) => el.key !== key)]);
+  };
+
+  const doCancel = () => {
+    setShowDeleteDialog(false);
+  };
+
+  const doDelete = () => {
+    setShowDeleteDialog(false);
+    const connectorService = Services.getConnectorService();
+    connectorService
+      .deleteConnector(appLayoutContext.clusterId, connectorToDelete)
+      .then((cConnectors: any) => {
+        addAlert("success", 'Connector deleted successfully!');
+      })
+      .catch((err: React.SetStateAction<Error>) => {
+        addAlert("danger",'Connector deletion failed!', err?.message);
+      });
+  };
+
+  const showConfirmationDialog = () => {
+    setShowDeleteDialog(true);
+  };
 
   const createConnector = () => {
     const connectorNames = connectors.map( (conn) => {
@@ -53,7 +88,6 @@ export const ConnectorsTableComponent: React.FunctionComponent<IConnectorsTableC
     props.createConnectorCallback(connectorNames, props.clusterId);
   }
 
-
   const getConnectorsList = () =>{
     const connectorService = Services.getConnectorService();
     fetch_retry(connectorService.getConnectors, connectorService, [
@@ -61,19 +95,118 @@ export const ConnectorsTableComponent: React.FunctionComponent<IConnectorsTableC
     ])
       .then((cConnectors: Connector[]) => {
         setLoading(false);
-        setConnectors([...cConnectors]);
+        updateTableRows([...cConnectors]);
       })
       .catch((err: React.SetStateAction<Error>) => {
         setApiError(true);
         setErrorMsg(err);
       });
+  }
+
+  const getTaskStates = (conn: Connector) => {
+    const taskElements: any = [];
+    const statesMap = new Map(Object.entries(conn.taskStates));
+    statesMap.forEach((taskState: any, id: string) => {
+      taskElements.push(
+        <ConnectorTask
+          key={id}
+          task={taskState.taskStatus}
+          taskId={id}
+          errors={taskState.errors}
+        />
+      );
+    });
+    return taskElements;
+  };
+  
+  React.useEffect(() => {
+    const timeout = setTimeout(
+      removeAlert,
+      10 * 1000,
+      alerts[alerts.length - 1]?.key
+    );
+    return () => clearTimeout(timeout);
+  }, [alerts]);
+
+  React.useEffect(() => {
+    const getConnectorsInterval = setInterval(() => getConnectorsList(), 10000);
+    return () => clearInterval(getConnectorsInterval);;
+  },[]);
+
+  const columns = [
+    {
+      title: '',
+      columnTransforms: [cellWidth(10)]
+    }, 
+    { 
+      title: 'Name', 
+      columnTransforms: [cellWidth(40)]
+    }, 
+    { 
+      title: 'Status', 
+      columnTransforms: [cellWidth(20)]
+    }, 
+    { 
+      title: 'Tasks', 
+      columnTransforms: [cellWidth(30)]
     }
+  ];
+  
+  const updateTableRows = (conns: Connector[]) => {
+    setConnectors(conns);
 
-    React.useEffect(() => {
-      const getConnectorsInterval = setInterval(() => getConnectorsList(), 10000);
-      return () => clearInterval(getConnectorsInterval);;
-    },[]);
+    // Sort connectors by name for the table
+    const sortedConns: Connector[] = conns.sort((thisConn, thatConn) => {
+      return thisConn.name.localeCompare(thatConn.name);
+    });
 
+    // Create table rows
+    const rows: any[] = [];
+    for (const conn of sortedConns) {
+      const row = {
+        cells: [
+          {
+            title: 
+              <ConnectorIcon
+                connectorType={
+                  conn.connectorType === "PostgreSQL"
+                    ? ConnectorTypeId.POSTGRES
+                    : conn.connectorType
+                }
+                alt={conn.name}
+                width={40}
+                height={40}
+              />
+          },
+          {
+            title: <b data-testid={"connector-name"}>{conn.name}</b>,
+          },
+          {
+            title: <ConnectorStatus currentStatus={conn.connectorStatus} />,
+          },
+          {
+            title: getTaskStates(conn)
+          }
+        ],
+        connName: conn.name
+      };
+      rows.push(row);
+    }
+    setTableRows(rows);
+  };
+
+  const tableActionResolver = () => {
+    return [
+      {
+        title: 'Delete',
+        onClick: (event, rowId, rowData, extra) => {
+          setConnectorToDelete(rowData.connName);
+          showConfirmationDialog();
+        }
+      }
+    ];
+  }
+    
   return (
     <WithLoader
       error={apiError}
@@ -85,9 +218,27 @@ export const ConnectorsTableComponent: React.FunctionComponent<IConnectorsTableC
         <>
           {connectors.length > 0 ? (
             <>
+              <ConfirmationDialog
+                buttonStyle={ConfirmationButtonStyle.DANGER}
+                i18nCancelButtonText={"Cancel"}
+                i18nConfirmButtonText={"Delete"}
+                i18nConfirmationMessage={
+                  "The connector will be deleted, and cannot be undone.  Proceed?"
+                }
+                i18nTitle={"Delete connector"}
+                type={ConfirmationType.DANGER}
+                showDialog={showDeleteDialog}
+                onCancel={doCancel}
+                onConfirm={doDelete}
+              />
+              <ToastAlertComponent alerts={alerts} removeAlert={removeAlert} />
               <Flex className="connectors-page_toolbarFlex">
                 <FlexItem>
-                  {props.title ? <Title headingLevel={"h1"}>Connectors</Title> : ""}
+                  {props.title ? (
+                    <Title headingLevel={"h1"}>Connectors</Title>
+                  ) : (
+                    ""
+                  )}
                 </FlexItem>
                 <FlexItem>
                   <Button
@@ -99,19 +250,16 @@ export const ConnectorsTableComponent: React.FunctionComponent<IConnectorsTableC
                   </Button>
                 </FlexItem>
               </Flex>
-              <DataList aria-label={"connector list"} className="connectors-page_dataList">
-                {getSortedConnectors(connectors).map((conn, index) => {
-                  return (
-                    <ConnectorListItem
-                      key={index}
-                      name={conn.name}
-                      type={conn.connectorType}
-                      status={conn.connectorStatus}
-                      taskStates={conn.taskStates}
-                    />
-                  );
-                })}
-              </DataList>
+              <Table 
+                aria-label="Connector Table" 
+                className="connectors-page_dataTable"
+                cells={columns} 
+                rows={tableRows}
+                actionResolver={tableActionResolver}
+              >
+                <TableHeader />
+                <TableBody />
+              </Table>
             </>
           ) : (
             <EmptyState variant={EmptyStateVariant.large}>
