+++ ui/packages/ui/webpack.common.js
@@ -46,8 +46,7 @@ module.exports = {
     rules: [
       {
         test: /\.(tsx|ts|js)?$/,
-        //include: path.resolve(__dirname, 'src'),
-        exclude: /node_modules/,
+        include: path.resolve(__dirname, 'src'),
         use: [
           {
             loader: 'ts-loader',
@@ -124,6 +123,7 @@ module.exports = {
           path.resolve(__dirname, 'node_modules/patternfly'),
           path.resolve(__dirname, 'node_modules/@patternfly/patternfly/assets'),
           path.resolve(__dirname, 'node_modules/@patternfly/react-core/dist/styles/assets/images'),
+          path.resolve(__dirname, '../../node_modules/@patternfly/react-table/node_modules/@patternfly/react-styles/css/assets/images'),
           path.resolve(__dirname, 'node_modules/@patternfly/react-styles/css/assets/images')
         ],
         use: [
